Interaction base
Interaction logic implemented.
For now the only available interaction is sitting.
I added an align point to InteractionTarget and a finish point as the position where the character should move when interaction is finished (for example, when standing up from sitting), but for now these are unused.
Next - implement standing - consider also making it a command.